requirements/lista_v3.txt
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
ipython==8.5.0
matplotlib-inline==0.1.6
numpy==1.23.4
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.5.0
xlrd==2.0.1
xls2xlsx==0.2.0
xlwt==1.3.0
django==4.0.6
django-allauth
sphinx
myst-parser
furo
django-crispy-forms
mercadopago
django-phonenumber-field
phonenumbers
django-autocomplete-light
django-bootstrap4
crispy-bootstrap4
pytest-django
faker
django-dynamic-fixture
factory-boy
jwt
websockets